refactor: 도메인 별 폴더/계층 구분
