+++ src/main/java/com/sejong/elasticservice/postlike/consumer/PostlikeConsumer.java
@@ -1,8 +1,10 @@
-package com.sejong.elasticservice.postlike;
+package com.sejong.elasticservice.postlike.consumer;
 
-import com.sejong.elasticservice.common.name.GroupNames;
-import com.sejong.elasticservice.common.name.TopicNames;
-import com.sejong.elasticservice.project.ProjectElasticRepository;
+import com.sejong.elasticservice.common.constants.GroupNames;
+import com.sejong.elasticservice.common.constants.TopicNames;
+import com.sejong.elasticservice.postlike.domain.PostLikeEvent;
+import com.sejong.elasticservice.postlike.PostType;
+import com.sejong.elasticservice.project.repository.ProjectElasticRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.stereotype.Component;
@@ -20,7 +22,7 @@ public class PostlikeConsumer {
 
         PostLikeEvent event = PostLikeEvent.fromJson(message);
 
-        if(event.getPostType()==PostType.PROJECT){
+        if(event.getPostType()== PostType.PROJECT){
             projectElasticRepository.updateLikeCount(event.getPostId(), event.getLikeCount());
         }
         else if (event.getPostType()==PostType.ARTICLE){
