+++ src/main/java/com/sejong/elasticservice/postlike/domain/PostType.java
