+++ src/main/java/com/sejong/elasticservice/project/domain/ProjectDocument.java
@@ -2,8 +2,8 @@ package com.sejong.elasticservice.project.domain;
 
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import com.sejong.elasticservice.project.ProjectEvent;
-import com.sejong.elasticservice.project.ProjectStatus;
+import com.sejong.elasticservice.project.domain.ProjectEvent;
+import com.sejong.elasticservice.project.domain.ProjectStatus;
 import lombok.*;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.elasticsearch.annotations.*;
