+++ src/main/java/com/sejong/elasticservice/project/domain/ProjectStatus.java
@@ -1,4 +1,4 @@
-package com.sejong.elasticservice.project;
+package com.sejong.elasticservice.project.domain;
 
 
 public enum ProjectStatus {
