+++ src/main/java/com/sejong/elasticservice/project/domain/Type.java
