+++ src/main/java/com/sejong/elasticservice/view/domain/ViewEvent.java
@@ -1,9 +1,8 @@
-package com.sejong.elasticservice.postlike;
+package com.sejong.elasticservice.view.domain;
 
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.sejong.elasticservice.project.ProjectIndexEvent;
+import com.sejong.elasticservice.postlike.PostType;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -13,16 +12,15 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
-@JsonIgnoreProperties(ignoreUnknown=true)
-public class PostLikeEvent {
+public class ViewEvent {
     private Long postId;
     private PostType postType;
-    private Long likeCount;
+    private Long viewCount;
 
-    public static PostLikeEvent fromJson(String message) {
+    public static ViewEvent fromJson(String message) {
         try {
             ObjectMapper objectMapper = new ObjectMapper();
-            return objectMapper.readValue(message, PostLikeEvent.class);
+            return objectMapper.readValue(message, ViewEvent.class);
         } catch (JsonProcessingException e) {
             throw new RuntimeException("JSON 파싱 실패");
         }
